Add test.out to repo.
Code works now by means of repeated calls to the calc_next_lake_level method. I did it this way so I
can see the output line by line working in IPython. Could finish this up by just calling this method
for each set of points in the NINPUT.DAT file.

Mike, I've luckily only been working on this for a few minutes here and a few minutes there; feel free
to spiffy it up. I was going to play with Matplotlib to put together a quick comparison of the Fortran
output vs the Python output, but that will have to wait for another day.